Added Alerted Enemy Script
Enemies can now alert each other that they have spotted the player with
the template script.

EnemyInfo could be a basic script that stores information on the enemy
that needs to be accessed by other objects/enemies/player/etc. So far I
have the basic enemies accessing the EnemyInfo script to see if the
other enemies have "seen" the player or have been "alerted" of the
player's location.