Added Lose of Sight to Alerted groups
Not sure if there's a more efficient way to do this but at the moment,
each enemy that has the player in sight (and is not just alerted by the
others) reports to a parent script that contains the number of enemies
have the player in sight. The parent script then resets itself each
FixedUpdate to count again. The children/enemies use the script to see
if any of them can see the player.

What this means is we have kind of an Assassin's Creed alert system.
After one enemy 'sees' the player and 'alerts' another, all the alerted
enemies (+ the one that saw the player in the first place) will continue
to pursue the player until he/she has completely LoS'ed ALL the enemies.