Added Enemy Alert System with Loss of Sight
Sorry, I think I messed up the save earlier. Here's the game with a
working Alert System.

EnemyInfo script is used to access information on the enemy by other
objects/enemies/player. So far it's used to alert other enemies and to
report to the 'Global' enemy script that records how many enemies can
see the player.

With how the alert system is now, all the alerted enemies + the enemy
that spotted the player originally, will continue to pursue the player
until the player completely LoS's all enemies (or an enemy will stop
pursuing if there is no one 'alerted' next to it)