Fixed the alert system and minor clean up of scripts. It should work as intended now somewhat. The player still has to LoS all the enemies for all the Alerted enemies to stop chasing the player.
GlobalEnemyInfo: should be used as parent of enemies so they can all the 'alert' system and know when to stop chasing the player. Also, if you're testing the spawning script, you might also have to put it a child of the GlobalEnemyInfo script gameObject so that the spawned enemies can use the parent script. Haven't quite tested it yet.