Fixed hacked enemy animation issues.
The player should no longer revert to the old sprite (hopefully). The
problem was earlier there was a case statement that took the names of
the enemies (Ex. rangedEnemy) as a way to set the EnemyType for the
animation. However, some of the enemies are actually called
"rangedEnemy(Clone)" or "basicEnemy(Clone)" instead of just
"rangedEnemy" or "basicEnemy". I changed the code so that the player
instead reads from the ENEMYINFO script and gets the ENEMYTYPE from
there.