Player can now use ranged enemy attack.
For now the player attacks in intervals after it hacks the ranged enemy.
This can be changed to mouse clicking.

Moved scripts for takeDamage for enemies down to ObjectTag because
that's where the tags for Enemies are and SendMessage applies directly
to the scripts where the actual tag is.

Basic Enemies are killable with the bullets. Ranged enemies don't have
death set for them yet.