Tiled based Gate placement
The map to the left is an example of using Tiled2Unity and some script
to place the gate connectors. The connectors for the Test map on the
left are not visible because they don't exist until the game starts and
the script creates them.

Scale should be set correctly. All gate placement can be move slightly
if they seem off. Also, I didn't need to hand draw any of the colliders
outside of Tiled if that was an issue.

Also weird thing where you cannot see the Connectors in Scene view for
the Lvl 1 but you can see them in the game view. They should still
function the same as before.