Temporary fix for teleporting/hacking issue
For some reason setting inHackingAnim = true with the Connectors make
the players when hit by the bullets. The problem doesn't seem to happen
though when it's left alone on Enemies. Though that might just be
because I can't replicate the effect on enemies. Maybe the player
doesn't fly because the Enemy Hack moves the player to the Enemy's
position right before they fly?

Either way, right now we can move while hacking Gate Connectors.