Fixed ranged hacked (programming note)
If you want to be able to do something to ALL enemies (not just the
basic or not just the ranged), you have to access their shared script
and not their individual script. For example, doing anything to
<TemplateEnemyAI> only applies to the Basic Enemies and not the ranged.
To access the Ranged enemies information (specifically) you'd have to
access <RangedEnemyAI>.

But there are cases when we want to do the same thing to ALL enemy types
which is why we need EnemyInfo which is a script that should be on all
Enemies. So when you want to perform an action across all enemy types,
access and change their <EnemyInfo> script. Then have something in their
AI script look at EnemyInfo for the information it needs.