Programmers' note. Also Fixed looping death
...for ranged Enemy. Also ranged enemies can die to ranged enemy bullets
already. So they are able to receive sendmessage (takeDamage)

The "takeDamage" script had to be moved to a GameObject with the actual
tag "Enemy". That means the script exists in the ranged enemy's
"ObjectTag".

This was because when you SEND MESSAGE to a gameobject with a specific
tag, it'll only SEND MESSAGE to the gameobject with that tag. That means
when the player did SEND MESSAGE ( takeDamage) to the ranged enemy, it
was sending that message to its child OBJECT TAG which originally had no
such function until the takeDamage() was moved to be a part of the
component script BasicSMessage in Range Enemy's game object "Object Tag"